% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot1g.image.R
\name{boot1g.image}
\alias{boot1g.image}
\title{Select triangulation for one sample SCC construction using bootstrap}
\usage{
boot1g.image(Y, Z, d.band, r, B.est, Q2.est, K.est, B.bands, Q2.bands,
  K.bands, ind.inside, lambda, nboot = 50, alpha0, adjust.sigma = TRUE)
}
\description{
Select triangulation for one sample SCC construction using bootstrap
}
