% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2g.image.R
\name{boot2g.image}
\alias{boot2g.image}
\title{Select triangulation for two sample SCC construction}
\usage{
boot2g.image(Ya, Yb, Z, d.band, r, B.est.a, Q2.est.a, K.est.a, B.est.b,
  Q2.est.b, K.est.b, tri.boot = NULL, B.bands, Q2.bands, K.bands,
  ind.inside, lambda, nboot = 50, alpha0 = 0.05, adjust.sigma = TRUE)
}
\description{
Select triangulation for two sample SCC construction
}
