% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1g.image.R
\name{data1g.image}
\alias{data1g.image}
\title{Generate simulation examples: one sample case}
\usage{
data1g.image(n, Z, ind.inside, mu.func = c(1, 2, 3, 4),
  noise.type = c("Func", "Const"), lam1, lam2, iter = 2019)
}
\arguments{
\item{n}{number of images in the simulated data.}

\item{Z}{a 2-column matrix specifying locations of information.}

\item{ind.inside}{a vector of indices specifying which locations in matrix Z fall inside
the irregular domain.}

\item{mu.func}{a integer in \code{c(1,2,3,4)} that specifies what mean function is used for
data generation. Value \code{1,2,3,4} correspond to quadratic, exponential, cubic and sine
functions described in the paper respectively.}

\item{noise.type}{'Func'/'Const' indicating heterogeneous or homogeneous variance for measurement error.}

\item{lam1, lam2}{the eigenvalues used to adjust subject-level variation in simulation data.}

\item{iter}{random seed.}
}
\description{
The function generates one-sample simulation data using four different mean functions
and scheme described in the paper.
}
\details{
This R package is the implementation program for manuscript entitled "Simultaneous Confidence Corridors for Mean Functions in Functional Data Analysis of Imaging Data" by Yueying Wang, Guannan Wang, Li Wang and R. Todd Ogden.
}
