% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2g.image.R
\name{data2g.image}
\alias{data2g.image}
\title{Generate simulation examples: two sample case}
\usage{
data2g.image(na, nb, Z, ind.inside, mu1.func, noise.type = c("Func",
  "Const"), lam1, lam2, delta, iter = 2019)
}
\arguments{
\item{na}{number of images/subjects in the first sample.}

\item{nb}{number of images/subjects in the second sample.}

\item{Z}{a 2-column matrix specifying locations of information.}

\item{ind.inside}{a vector of indices specifying which locations in matrix Z fall inside
the irregular domain.}

\item{mu1.func}{a integer in \code{c(1,2,3,4)} that specifies what mean function is used for
data generation of the first sample. Value \code{1,2,3,4} correspond to quadratic, exponential, cubic and sine
functions described in the paper respectively.}

\item{noise.type}{'Func'/'Const' indicating heterogeneous or homogeneous variance for measurement error.}

\item{lam1, lam2}{the eigenvalues used to adjust subject-level variation in simulation data.}

\item{delta}{a parameter indicating the scale of difference between two samples' mean functions.}

\item{iter}{random seed.}
}
\description{
The function generates two-sample simulation data using four different mean functions
and eigenfunctions described in the paper.
}
\details{
This R package is the implementation program for manuscript entitled "Simultaneous Confidence Corridors for Mean Functions in Functional Data Analysis of Imaging Data" by Yueying Wang, Guannan Wang, Li Wang and R. Todd Ogden.
}
