% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.image.R
\name{plot.image}
\alias{plot.image}
\title{Plot image with triangulation}
\usage{
\method{plot}{image}(mfit)
}
\arguments{
\item{mfit}{an "image" object returned from either \code{fit.image} or \code{scc.image}.}
}
\description{
The plot method for class 'image'. The function plots a series of images of triangulations used for mean estimation, estimated mean functions and simultaneous confidence corridors.
}
\details{
This R package is the implementation program for manuscript entitled "Simultaneous Confidence Corridors for Mean Functions in Functional Data Analysis of Imaging Data" by Yueying Wang, Guannan Wang, Li Wang and R. Todd Ogden.
}
