% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc1g.image.R
\name{scc1g.image}
\alias{scc1g.image}
\title{Simultaneous confidence corridors for mean function of imaging data: one sample case}
\usage{
scc1g.image(Y, Z, Z.band, d.est, d.band, r, B.est, Q2.est, K.est, B.band,
  Q2.band, K.band, ind.inside, B.cover1, B.cover2, ind.inside.cover,
  penalty = FALSE, lambda, alpha.grid = c(0.1, 0.05, 0.01),
  adjust.sigma = TRUE)
}
\description{
Simultaneous confidence corridors for mean function of imaging data: one sample case
}
