% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc2g.image.R
\name{scc2g.image}
\alias{scc2g.image}
\title{Simultaneous confidence corridors for mean function of imaging data: two sample case}
\usage{
scc2g.image(Ya, Yb, Z, Z.band, d.est, d.band, r, B.est.a, Q2.est.a,
  K.est.a, B.est.b, Q2.est.b, K.est.b, B.band.a, Q2.band.a, K.band.a,
  B.band.b, Q2.band.b, K.band.b, ind.inside, B.cover1.a, B.cover2.a,
  B.cover1.b, B.cover2.b, ind.inside.cover, penalty = TRUE, lambda,
  alpha.grid = c(0.1, 0.05, 0.01), adjust.sigma = TRUE)
}
\description{
Simultaneous confidence corridors for mean function of imaging data: two sample case
}
